pyproject.toml
[project]
name = "pralapin-backend"
version = "0.1.0"
description = "Pralapin SMS - FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "beanie>=1.25.0",
    "motor>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "firebase-admin>=6.4.0",
    "weasyprint>=60.0",
    "reportlab>=4.0.0",
    "httpx>=0.26.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
dev-dependencies = []
